Changes:
Removed unnescessary files in the opmode folder and in the module folder. Updated opmode to include one module, inside of which has example code for an to show format to set up motors.
Rest of code left alone - extra examples.

Rest to go : simply add in the motors and servos in 'GolfMotorServo' - then run opmode and it's done.
- create config on driver station, putting motors in correct ports.